requirements.txt
streamlit
pygithub
sentence-transformers
faiss-cpu
python-dotenv
google-generativeai
